+++ src/algorithms/ImgRegression.h
@@ -30,12 +30,12 @@ namespace imgregression
   public:
     ImgRegression(void);
     ~ImgRegression(void);
-    double getRMSE(ImgRaster& imgReader1, ImgRaster& imgReader2, double &c0, double &c1, unsigned short b1=0, unsigned short b2=0, short verbose=0) const;
-    double getRMSE(ImgRaster& imgReader, unsigned short b1, unsigned short b2, double& c0, double& c1, short verbose=0) const;
-    double getR2(ImgRaster& imgReader1, ImgRaster& imgReader2, double &c0, double &c1, unsigned short b1=0, unsigned short b2=0, short verbose=0) const;
-    double pgetR2(ImgRaster& imgReader1, ImgRaster& imgReader2, double& c0, double& c1, unsigned short band1, unsigned short band2, short verbose=0) const;
-    double getR2(ImgRaster& imgReader, unsigned short b1, unsigned short b2, double& c0, double& c1, short verbose=0) const;
-    double pgetR2(ImgRaster& imgReader, unsigned short band1, unsigned short band2, double& c0, double& c1, short verbose=0) const;
+    double getRMSE(ImgRaster& imgReader1, ImgRaster& imgReader2, double &c0, double &c1, unsigned int b1=0, unsigned int b2=0, short verbose=0) const;
+    double getRMSE(ImgRaster& imgReader, unsigned int b1, unsigned int b2, double& c0, double& c1, short verbose=0) const;
+    double getR2(ImgRaster& imgReader1, ImgRaster& imgReader2, double &c0, double &c1, unsigned int b1=0, unsigned int b2=0, short verbose=0) const;
+    double pgetR2(ImgRaster& imgReader1, ImgRaster& imgReader2, double& c0, double& c1, unsigned int band1, unsigned int band2, short verbose=0) const;
+    double getR2(ImgRaster& imgReader, unsigned int b1, unsigned int b2, double& c0, double& c1, short verbose=0) const;
+    double pgetR2(ImgRaster& imgReader, unsigned int band1, unsigned int band2, double& c0, double& c1, short verbose=0) const;
 
     void setThreshold(double theThreshold){m_threshold=theThreshold;};
     void setDown(int theDown){m_down=theDown;};
